requirements.txt
python-dotenv
scrapy
elasticsearch=8.15.1